refactor(job): job detail
